Reformat FOSS License Tool
