+++ conf/layer.conf
@@ -15,6 +15,4 @@ LAYERSERIES_COMPAT_meta-foss-license-checker = "scarthgap"
 INHIBIT_SYSROOT_STRIP = "1"
 LAYERDIR_WS = "${LAYERDIR}"
 
-addpylib ${LAYERDIR}/lib static
-addpylib ${LAYERDIR}/lib dynamic
-addpylib ${LAYERDIR}/lib foss
+addpylib ${LAYERDIR}/src lib
